some 64 bits warning removed

+++ executables/MISC/TestMyZLib/src/main.cpp
@@ -42,7 +42,7 @@ chaos::Buffer<char> GetIpsumBuffer()
 {
 	static char * ipsum = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
 
-	int length = strlen(ipsum) + 1;
+	int length = (int)strlen(ipsum) + 1;
 
 	chaos::SparseWriteBuffer<> writer(512);
